+++ ICD.Connect.Krang/Properties/AssemblyInfo.cs
@@ -5,3 +5,8 @@ using System.Reflection;
 [assembly: AssemblyProduct("ICD.Connect.Krang")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
 [assembly: AssemblyVersion("3.1.0.0")]
+
+[assembly: System.Reflection.AssemblyInformationalVersion("2.0.0")]
+#if SIMPLSHARP
+[assembly: Crestron.SimplSharp.Reflection.AssemblyInformationalVersion("2.0.0")]
+#endif