+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Added VersionInfo console command to KrangBootstrap
+ - Added informational version info to AssemblyInfo
 
 ## [3.1.0] - 2018-04-15